Update new.md with 9.0 SDK Release Notes
Updated new.md with 9.0 SDK Release Notes content from the MyGeotab Blog at https://www.geotab.com/blog/mygeotab-updates-9/

Note: Wrote content / created this Commit to demonstrate Technical Writing capability for Markdown and potential for streamlining shared processes. For 10.0 and going forward, we can submit this to the repo as an outcome of publication so that the Public and Dev site details are in sync? @beefo let me know your thoughts on this approach and happy to Meet to discuss further.
Co-Authored-By: Steve Hansen <beefo@hotmail.com>